Make debouncing tests reliable (#65)
Previously, these tests stubbed the `_fork` method of the instance under
test. The problem was that this method was only ever called using
`setImmediate`. That made it difficult to reliably test the timing
sensible behavior that was to be tested by these tests.

That's why the new test stub `_send` the method one step higher in the
stack. The exact nature of the batches is not tested anymore. Instead
it is only verified that the expected number of batches has been sent.

Fixes #64